syatic ==> templates   경로설정 변경
