+++ .github/JoinProject/src/main/resources/application.properties
@@ -1,5 +1,6 @@
 #http port
 server.port=8090
+server.servlet.context-path=/app
  
 #db connection(Oracle)
 spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
