--- .github/JoinProject/src/main/resources/static/Login.html
@@ -1,244 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Elementery Comunity</title>
-
-
-	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-	<script defer src="../modal.js"></script>
-    <script defer src="../schoolSearch.js"></script>
-    
-    <link rel="stylesheet" href="../LoginStyle.css" type="text/css">
-    <link rel="stylesheet" href="../modal.css">
-    <link rel="stylesheet" href="../schSearch.css">
-    
-</head>
-
-  
-<body>
-
-	
-				<!-- 사이트 중앙 헤더 텍스트 -->
-			   <h2>엄             준             식</h2>
-	
-	
-					<!-- 회원가입 폼 -->
-	
-<div class="container" id="container">
-  <div class="form-container sign-up-container" id="app">
-    <form id="frm">
-    
-      <h1>Create Account</h1>
-      <!-- 
-      <div class="social-container">
-        <a href="#" class="social"><i class="fab fa-facebook-f"></i></a>
-        <a href="#" class="social"><i class="fab fa-google-plus-g"></i></a>
-        <a href="#" class="social"><i class="fab fa-linkedin-in"></i></a>
-      </div>
-       -->
-      
-      <span>or use your email for registration</span>
-      
-      
-      	<input type="text" id="accountid" name="id" placeholder="아이디" />
-      	<span id="idcheck"></span>
-	  	<button type="button" id="idcheckBtn">중복확인</button>
-	
-      <input type="password" id="accountpwd" name="pwd" placeholder="비밀번호" />
-      
-      <input type="email" id="accountemail" name="email" placeholder="Email" />
-      
-      <input type="text" id="accounname" name="name" placeholder="이름" />
-      
-      
-				
-      <!--  --><input type="text" placeholder="학교선택 (추가예정)" />
-      
-      
-      
-      
-      
-      
-      <button type="button" id="account">Sign Up</button>
-  
-      
-    </form>
-  </div>
-  
-  
-  					<!-- 로그인 폼 -->
-  
-  <div class="form-container sign-in-container" id="app">
-    <form action="#">
-    
-    
-      <h1>Sign in</h1>
-      <div class="social-container">
-        <a href="#" class="social"><i class="fab fa-facebook-f"></i></a>
-        <a href="#" class="social"><i class="fab fa-google-plus-g"></i></a>
-        <a href="#" class="social"><i class="fab fa-linkedin-in"></i></a>
-      </div>
-      
-      
-      <span>use your account</span>
-      <input type="text" id="id" placeholder="아이디입력" />
-      <input type="password" id="pwd" placeholder="비밀번호 입력" />
-     
-      <a href="account.html">Forgot your password?</a>
-      
-      <button type="button" id="login">Sign In</button>
-      
-      
-    </form>
-  </div>
-  <div class="overlay-container">
-    <div class="overlay">
-      <div class="overlay-panel overlay-left">
-        <h1>Welcome Back!</h1>
-        <p>To keep connected with us please login with your personal info</p>
-        <button class="ghost" id="signIn">Sign In</button>
-      </div>
-      <div class="overlay-panel overlay-right">
-        <h1>Hello, Friend!</h1>
-        <p>Enter your personal details and start journey with us</p>
-        <button class="ghost" id="signUp">Sign Up</button>
-      </div>
-    </div>
-  </div>
-</div>
-
-
-<script type="text/javascript">
-
-//#################### 로그인 ############################
-
-$(document).ready(function(){
-	
-	$("#login").click(function(){ //로그인 버튼 누를시 실행되는 함수
-		
-		$.ajax({
-			url:"http://localhost:8090/login",
-			type:"post",
-			data:{ id:$("#id").val(), pwd:$("#pwd").val() }, //id란과 pwd에 입력된 Data 서버로 전달
-			
-			//백엔드에서 넘겨주는 데이터가 dto라는 객체이므로 json으로 받는다
-			success:function(json){
-			//	alert('success');
-			//	alert(JSON.stringify(json));
-				if(json==""){
-					alert("id나 비밀번호를 확인하세요.")
-					$("#id").val("");
-					$("#pwd").val("")
-				}else{
-					sessionStorage.setItem("login",JSON.stringify(json));
-					alert(json.name+"님 환영합니다.")
-					location.href="bbslist.html"
-				}
-				
-			},
-			error:function(){
-				alert('error');
-			}
-			
-		});
-		
-		
-	});
-	
-	
-});
-
-
-// #################### 회원가입 ############################
-
-$(document).ready(function() {
-	
-	$("#idcheckBtn").click(function() {
-		$.ajax({
-			url:"http://localhost:8090/getId",
-			type:"post",
-			data:{"id":$("#accountid").val() },
-			success:function( msg ){
-				//alert('success')
-				//alert(msg);
-				if(msg.trim()=='OK'){
-					$("#idcheck").html("이 아이디는 사용 가능합니다.");
-					$("#accountid").val( $("#accountid").val());
-				}else{
-					$("#idcheck").html("이 아이디는 사용중입니다.");
-					$("#accountid").val("");//삭제
-					$("#accountid").focus(); //다시 입력하라고 커서가 뜬다.
-				}
-			},
-			error:function(){
-				alert('error');
-			}
-		});
-		
-	});
-	
-	$("#account").on("click",function(){
-		
-		let params = $("#frm").serialize()
-		
-		$.ajax({
-			url:"http://localhost:8090/addmember",
-			type:"post",
-			//data:{id:$("#id").val(), pwd:$("#pwd").val(), name:$("name").val(), email:$("email").val() }, //id로 보낸것
-			data:params,//name으로 보낸것
-			success:function(data){
-				//alert('success')
-				if(data=="YES"){
-					alert('가입성공');
-					location.href="login.html";
-				}else{
-					alert('가입실패');
-				}
-			},
-			error:function(){
-				alert('error');
-			}
-		});
-		
-	});
-	
-	
-	/*
-	
-	//#################### 정규식을 이용한 이메일 형식 체크 ############################
-
-	let input = document.getElementById("accountemail").value
-	let email_format = /^([0-9a-zA-Z_\.-]+)@([0-9a-zA-Z_-]+)(\.[0-9a-zA-Z_-]+){1,2}$/;
-
-	if (email_format.test(input)) {
-	    alert("이메일 입력 성공!")
-	} else {
-	    alert("잘못된 이메일 형식입니다");
-	    document.getElementById("accountemail").value = "";
-	    inputform.em.focus();
-	}
-	
-	*/
-});
-
-
-	
-	
-
-//#################### 템플릿 ############################
-const signUpButton = document.getElementById('signUp');
-const signInButton = document.getElementById('signIn');
-const container = document.getElementById('container');
-
-signUpButton.addEventListener('click', () => {
-		container.classList.add("right-panel-active");
-});
-
-signInButton.addEventListener('click', () => {
-		container.classList.remove("right-panel-active");
-});
-</script>
-
-</body>
-</html>
