--- .github/JoinProject/src/main/resources/static/account.html
@@ -1,134 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-
-<style type="text/css">
-#aap{
-	margin:auto;
-	margin-top:60px;
-	width: 60%;
-	border: 3px solid #8ac007;
-	padding: 10px;
-}
-p{
-font-size:8px;
-}
-
-</style>
-
-</head>
-<body>
-
-<h1>회원가입</h1>
-<div id="app">
-<form id="frm">
-
-<table border="1">
-	<tr>
-			<td>아이디</td>
-			<td>
-			<input type="text" id="userid" placeholder="아이디입력" size="20">
-			<p id="idcheck"></p>
-			<button type="button" id="idcheckBtn">중복확인</button>
-			</td>
-	</tr>
-	
-	<tr>
-			<td>사용 아이디</td>
-			<td>
-			<input type="text" id="id" name="id" size="20" readonly="readonly">
-			</td>
-	</tr>
-	
-	
-	<tr>
-			<td>패스워드</td>
-			<td>
-			<input type="text" id="pwd" name="pwd" placeholder="패스워드 입력" size="20">
-			</td>
-	</tr>	
-		<tr>
-			<td>이름</td>
-			<td>
-			<input type="text" id="name" name="name" placeholder="이름 입력" size="20">
-			</td>
-	</tr>	
-	
-		<tr>
-			<td>이메일</td>
-			<td>
-			<input type="text" id="email" name="email" placeholder="이메일 입력" size="20">
-			</td>
-	</tr>	
-	
-	<tr>
-		<td colspan="2">
-		<button type="button" id="account">회원가입</button>
-		</td>
-	</tr>
-	</table>
-	</form>
-	</div>
-	
-	<script type="text/javascript">
-	$(document).ready(function() {
-		
-		$("#idcheckBtn").click(function() {
-			$.ajax({
-				url:"http://localhost:8090/getId",
-				type:"post",
-				data:{"id":$("#userid").val() },
-				success:function( msg ){
-					//alert('success')
-					//alert(msg);
-					if(msg.trim()=='OK'){
-						$("#idcheck").html("이 아이디는 사용 가능합니다.");
-						$("#id").val( $("#userid").val());
-					}else{
-						$("#idcheck").html("이 아이디는 사용중입니다.");
-						$("#userid").val("");//삭제
-						$("#userid").focus(); //다시 입력하라고 커서가 뜬다.
-					}
-				},
-				error:function(){
-					alert('error');
-				}
-			});
-			
-		});
-		
-		$("#account").on("click",function(){
-			
-			let params = $("#frm").serialize()
-			
-			$.ajax({
-				url:"http://localhost:8090/addmember",
-				type:"post",
-				//data:{id:$("#id").val(), pwd:$("#pwd").val(), name:$("name").val(), email:$("email").val() }, //id로 보낸것
-				data:params,//name으로 보낸것
-				success:function(data){
-					//alert('success')
-					if(data=="YES"){
-						alert('가입성공');
-						location.href="login.html";
-					}else{
-						alert('가입실패');
-					}
-				},
-				error:function(){
-					alert('error');
-				}
-			});
-			
-		});
-		
-	});
-	
-	</script>
-	
-</body>
-</html>
