+++ .github/JoinProject/src/main/resources/static/js/modal.js
@@ -1,0 +1,9 @@
+$('#modalBtn').click(function(e){
+			e.preventDefault();
+			$('#schModal').modal("show");
+		});
+
+$('#closeModal').click(function(e){
+	e.preventDefault();
+	$('#schModal').modal("hide");
+});
