--- .github/JoinProject/src/main/resources/static/login2.html
@@ -1,89 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-
-
-<style type="text/css">
-#aap{
-	margin:auto;
-	margin-top:60px;
-	width: 60%;
-	border: 3px solid #8ac007;
-	padding: 10px;
-}
-
-</style>
-
-</head>
-<body>
-
-<h1>Login</h1>
-<div id="app">
-<table border="1">
-	<tr>
-			<td>아이디</td>
-			<td>
-			<input type="text" id="id" placeholder="아이디입력" size="20">
-			</td>
-	</tr>
-	<tr>
-			<td>패스워드</td>
-			<td>
-			<input type="text" id="pwd" placeholder="패스워드 입력" size="20">
-			</td>
-	</tr>
-	<tr>
-		<td colspan="2">
-		<button type="button" id="login">log-in</button>
-		<a href="account.html">회원가입</a>
-		</td>
-	</tr>
-
-</table>
-
-</div>
-
-<script type="text/javascript">
-$(document).ready(function(){
-	
-	$("#login").click(function(){ //로그인 버튼 누를시 실행되는 함수
-		
-		$.ajax({
-			url:"http://localhost:8090/login",
-			type:"post",
-			data:{ id:$("#id").val(), pwd:$("#pwd").val() }, //id란과 pwd에 입력된 Data 서버로 전달
-			
-			//백엔드에서 넘겨주는 데이터가 dto라는 객체이므로 json으로 받는다
-			success:function(json){
-			//	alert('success');
-			//	alert(JSON.stringify(json));
-				if(json==""){
-					alert("id나 비밀번호를 확인하세요.")
-					$("#id").val("");
-					$("#pwd").val("")
-				}else{
-					sessionStorage.setItem("login",JSON.stringify(json));
-					alert(json.name+"님 환영합니다.")
-					location.href="bbslist.html"
-				}
-				
-			},
-			error:function(){
-				alert('error');
-			}
-			
-		});
-		
-		
-	});
-	
-	
-});
-</script>
-
-</body>
-</html>
