--- .github/JoinProject/src/main/resources/static/school.html
@@ -1,43 +1,0 @@
-<!-- option => style="display:none;" -->
-				<!-- 시도교육청 코드 -->
-				<input type="text" id="ATPT_OFCDC_SC_CODE_SIGN" size="20" value="">
-				<!-- 시도교육청 이름 -->
-				<input type="text" id="ATPT_OFCDC_SC_NM_SIGN" size="20" value="">
-				<!-- 학교 코드 -->
-				<input type="text" id="SD_SCHUL_CODE_SIGN" size="20" value="">
-				<!-- 학교 이름 -->
-				<input type="text" id="SCHUL_NM_SIGN" size="20" value="">
-				<button class="modal open-button">검색</button>
-
-				<div class="modal-box">
-					<dialog class="modal" id="modal">
-					<body class="modal_body">
-						<button style="float: right;" class="button close-button">x</button>
-						<h2>학교 찾기</h2>
-						<select id="ATPT_OFCDC_SC_CODE">
-							<option value="B10">서울특별시교육청</option>
-							<option value="C10">부산광역시교육청</option>
-							<option value="D10">대구광역시교육청</option>
-							<option value="E10">인천광역시교육청</option>
-							<option value="F10">광주광역시교육청</option>
-							<option value="G10">대전광역시교육청</option>
-							<option value="H10">울산광역시교육청</option>
-							<option value="I10">세종특별시교육청</option>
-							<option value="J10">경기도교육청</option>
-							<option value="K10">강원특별자치도교육청</option>
-							<option value="M10">충청북도교육청</option>
-							<option value="N10">충청남도교육청</option>
-							<option value="P10">전라북도교육청</option>
-							<option value="Q10">전라남도교육청</option>
-							<option value="R10">경상북도교육청</option>
-							<option value="S10">경상남도교육청</option>
-							<option value="T10">제주특별자치도교육청</option>
-						</select>
-						<input type="text" id="SCHUL_NM" size="20">
-						<a id="sheet-search-button" href="#" class="btn_A">검색</a>
-
-						<div class="SCH_NM" id="SCH_Name"></div>
-					</body>
-					<button class="button close-button">x</button>
-					</dialog>
-				</div>
