+++ .github/JoinProject/src/main/resources/templates/Login.html
@@ -5,13 +5,14 @@
 <title>Elementery Comunity</title>
 
 
-	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-	<script defer src="../modal.js"></script>
-    <script defer src="../schoolSearch.js"></script>
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
+	<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
+	<script defer src="./js/modal.js"></script>
+    <script defer src="./js/schoolSearch.js"></script>
     
-    <link rel="stylesheet" href="../LoginStyle.css" type="text/css">
-    <link rel="stylesheet" href="../modal.css">
-    <link rel="stylesheet" href="../schSearch.css">
+    <link rel="stylesheet" href="./css/LoginStyle.css" type="text/css">
+    <link rel="stylesheet" href="./css/modal.css">
+    <link rel="stylesheet" href="./css/schSearch.css">
     
 </head>
 
@@ -50,12 +51,60 @@
       <input type="email" id="accountemail" name="email" placeholder="Email" />
       
       <input type="text" id="accounname" name="name" placeholder="이름" />
-				
-      <input type="text" placeholder="학교선택 (추가예정)" />
+      
+      <!-- 모달@ㅉㄴㅃ@@@@@@@@@@@@@@@@@ㅃ@ㅃ@ㅃㅃ@ㅃ@ㅃ@ㅃ@ㅃ@ㅃ@ㅃ@ㅃㅃ@@ㅃㅃ@@ㅃ@@@@ -->
+      
+     <input type="text" id="ATPT_OFCDC_SC_CODE_SIGN" size="20" value="">
+<!-- 시도교육청 이름 -->
+<input type="text" id="ATPT_OFCDC_SC_NM_SIGN" size="20" value="">
+<!-- 학교 코드 -->
+<input type="text" id="SD_SCHUL_CODE_SIGN" size="20" value="">
+<!-- 학교 이름 -->
+<input type="text" id="SCHUL_NM_SIGN" size="20" value="">
+   <button id="modalBtn" class="btn">모달 테스트</button>
+
+<div>dd?</div>
+
+<div class = "modal-box">
+<dialog class="modal" id="schModal">
+   <body class="modal_body">
+  <button style="float:right;" id="closeModal" class="btn close-button">x</button>
+  <h2>학교 찾기</h2>
+  <select id="ATPT_OFCDC_SC_CODE">
+  <option value = "B10">서울특별시교육청</option>
+  <option value = "C10">부산광역시교육청</option>
+  <option value = "D10">대구광역시교육청</option>
+  <option value = "E10">인천광역시교육청</option>
+  <option value = "F10">광주광역시교육청</option>
+  <option value = "G10">대전광역시교육청</option>
+  <option value = "H10">울산광역시교육청</option>
+  <option value = "I10">세종특별시교육청</option>
+  <option value = "J10">경기도교육청</option>
+  <option value = "K10">강원특별자치도교육청</option>
+  <option value = "M10">충청북도교육청</option>
+  <option value = "N10">충청남도교육청</option>
+  <option value = "P10">전라북도교육청</option>
+  <option value = "Q10">전라남도교육청</option>
+  <option value = "R10">경상북도교육청</option>
+  <option value = "S10">경상남도교육청</option>
+  <option value = "T10">제주특별자치도교육청</option>
+  </select>
+  <input type="text" id="SCHUL_NM" size="20">
+  <a id="sheet-search-button" href="#" class="btn_A">검색</a>
+  
+  <div class="SCH_NM" id="SCH_Name"></div>
+</body>
+   <button class="button close-button">x</button>
+</dialog>
+</div>
+      
+      
+      <!-- 모달@ㅉㄴㅃ@@@@@@@@@@@@@@@@@ㅃ@ㅃ@ㅃㅃ@ㅃ@ㅃ@ㅃ@ㅃ@ㅃ@ㅃ@ㅃㅃ@@ㅃㅃ@@ㅃ@@@@ -->
       
       
       
       
+      <!--  -->
       
       
       <button type="button" id="account">Sign Up</button>
@@ -116,7 +165,7 @@ $(document).ready(function(){
 	$("#login").click(function(){ //로그인 버튼 누를시 실행되는 함수
 		
 		$.ajax({
-			url:"http://localhost:8090/login",
+			url:"http://localhost:8090/app/login.html",
 			type:"post",
 			data:{ id:$("#id").val(), pwd:$("#pwd").val() }, //id란과 pwd에 입력된 Data 서버로 전달
 			
@@ -154,7 +203,7 @@ $(document).ready(function() {
 	
 	$("#idcheckBtn").click(function() {
 		$.ajax({
-			url:"http://localhost:8090/getId",
+			url:"http://localhost:8090/app/getId",
 			type:"post",
 			data:{"id":$("#accountid").val() },
 			success:function( msg ){
@@ -181,7 +230,7 @@ $(document).ready(function() {
 		let params = $("#frm").serialize()
 		
 		$.ajax({
-			url:"http://localhost:8090/addmember",
+			url:"http://localhost:8090/app/addmember",
 			type:"post",
 			//data:{id:$("#id").val(), pwd:$("#pwd").val(), name:$("name").val(), email:$("email").val() }, //id로 보낸것
 			data:params,//name으로 보낸것
