--- .github/JoinProject/src/main/resources/templates/modal.css
@@ -1,8 +1,0 @@
-dialog{
-	max-height: 60%;
-	overflow-y:auto;
-}
-.modal_body{
-	max-height: 80%;
-	overflow-y:auto;
-}
